# perf: subgroup membership