--- a/ecc/bls12-377/g2.go
+++ b/ecc/bls12-377/g2.go
@@ -187,9 +187,14 @@ func (p *G2Affine) IsInfinity() bool {
 
 // IsOnCurve returns true if the affine point p in on the curve.
 func (p *G2Affine) IsOnCurve() bool {
-	var point G2Jac
-	point.FromAffine(p)
-	return point.IsOnCurve() // call this function to handle infinity point
+	if p.IsInfinity() {
+		return true
+	}
+	var left, right fptower.E2
+	left.Square(&p.Y)
+	right.Square(&p.X).Mul(&right, &p.X)
+	right.Add(&right, &bTwistCurveCoeff)
+	return left.Equal(&right)
 }
 
 // IsInSubGroup returns true if the affine point p is in the correct subgroup, false otherwise.
@@ -481,12 +486,14 @@ func (p *G2Jac) IsOnCurve() bool {
 // and https://eprint.iacr.org/2022/352.pdf, sec. 4.2
 // ψ(p) = [x₀]P
 func (p *G2Jac) IsInSubGroup() bool {
-	var res, tmp G2Jac
-	tmp.psi(p)
-	res.mulWindowed(p, &xGen).
-		SubAssign(&tmp)
+	if !p.IsOnCurve() {
+		return false
+	}
+	var res, img G2Jac
+	img.psi(p)
+	res.mulBySeed(p)
 
-	return res.IsOnCurve() && res.Z.IsZero()
+	return res.Equal(&img)
 }
 
 // mulWindowed computes the 2-bits windowed double-and-add scalar
@@ -521,6 +528,13 @@ func (p *G2Jac) mulWindowed(q *G2Jac, s *big.Int) *G2Jac {
 
 	return p
 
+}
+
+// mulBySeed multiplies the point q by the seed xGen in Jacobian coordinates
+// using an optimized addition chain.
+func (p *G2Jac) mulBySeed(q *G2Jac) *G2Jac {
+	p.mulWindowed(q, &xGen)
+	return p
 }
 
 // psi sets p to ψ(q) = u o π o u⁻¹ where u:E'→E is the isomorphism from the twist to the curve E and π is the Frobenius map.
@@ -627,8 +641,8 @@ func (p *G2Affine) ClearCofactor(a *G2Affine) *G2Affine {
 func (p *G2Jac) ClearCofactor(q *G2Jac) *G2Jac {
 	// https://eprint.iacr.org/2017/419.pdf, 4.1
 	var xg, xxg, res, t G2Jac
-	xg.mulWindowed(q, &xGen)
-	xxg.mulWindowed(&xg, &xGen)
+	xg.mulBySeed(q)
+	xxg.mulBySeed(&xg)
 
 	res.Set(&xxg).
 		SubAssign(&xg).