--- a/ecc/bw6-633/g1.go
+++ b/ecc/bw6-633/g1.go
@@ -181,9 +181,14 @@ func (p *G1Affine) IsInfinity() bool {
 
 // IsOnCurve returns true if the affine point p in on the curve.
 func (p *G1Affine) IsOnCurve() bool {
-	var point G1Jac
-	point.FromAffine(p)
-	return point.IsOnCurve() // call this function to handle infinity point
+	if p.IsInfinity() {
+		return true
+	}
+	var left, right fp.Element
+	left.Square(&p.Y)
+	right.Square(&p.X).Mul(&right, &p.X)
+	right.Add(&right, &bCurveCoeff)
+	return left.Equal(&right)
 }
 
 // IsInSubGroup returns true if the affine point p is in the correct subgroup, false otherwise.
@@ -476,19 +481,21 @@ func (p *G1Jac) IsOnCurve() bool {
 
 // 3r P = (x+1)ϕ(P) + (-x^5 + x⁴ + x)P
 func (p *G1Jac) IsInSubGroup() bool {
-
+	if !p.IsOnCurve() {
+		return false
+	}
 	var uP, u4P, u5P, q, r G1Jac
-	uP.mulWindowed(p, &xGen)
-	u4P.mulWindowed(&uP, &xGen).
-		mulWindowed(&u4P, &xGen).
-		mulWindowed(&u4P, &xGen)
-	u5P.mulWindowed(&u4P, &xGen)
+	uP.mulBySeed(p)
+	u4P.mulBySeed(&uP).
+		mulBySeed(&u4P).
+		mulBySeed(&u4P)
+	u5P.mulBySeed(&u4P)
 	q.Set(p).SubAssign(&uP)
 	r.phi(&q).SubAssign(&uP).
 		AddAssign(&u4P).
 		AddAssign(&u5P)
 
-	return r.IsOnCurve() && r.Z.IsZero()
+	return r.Z.IsZero()
 }
 
 // mulWindowed computes the 2-bits windowed double-and-add scalar
@@ -523,6 +530,13 @@ func (p *G1Jac) mulWindowed(q *G1Jac, s *big.Int) *G1Jac {
 
 	return p
 
+}
+
+// mulBySeed multiplies the point q by the seed xGen in Jacobian coordinates
+// using an optimized addition chain.
+func (p *G1Jac) mulBySeed(q *G1Jac) *G1Jac {
+	p.mulWindowed(q, &xGen)
+	return p
 }
 
 // phi sets p to ϕ(a) where ϕ: (x,y) → (w x,y),
@@ -629,7 +643,7 @@ func (p *G1Jac) ClearCofactor(q *G1Jac) *G1Jac {
 	ht.SetInt64(7)
 	v.Mul(&xGen, &xGen).Add(&v, &one).Mul(&v, &uPlusOne)
 
-	uP.mulWindowed(q, &xGen).Neg(&uP)
+	uP.mulBySeed(q).Neg(&uP)
 	vP.Set(q).SubAssign(&uP).
 		mulWindowed(&vP, &v)
 	wP.mulWindowed(&vP, &uMinusOne).Neg(&wP).