… ecc/bls12-381/g1.go
… ecc/bls12-381/g2.go
… internal/generator/ecc/template/point.go.tmpl